Refactored the components to extract the styles in separate files
