+++ src/components/CheckBoxWithLabel/CheckBoxWithLabelStyle.js
@@ -1,0 +1,14 @@
+import {StyleSheet} from 'react-native';
+import {colors} from '../../util/colors';
+
+export const styles = StyleSheet.create({
+  checkBoxText: {
+    alignSelf: 'center',
+    fontSize: 16,
+    color: colors.lightGray,
+  },
+  checkBoxContainer: {
+    alignContent: 'center',
+    flexDirection: 'row',
+  },
+});
