+++ src/components/TextBox/TextBoxStyle.js
@@ -1,0 +1,11 @@
+import {StyleSheet} from 'react-native';
+import {colors} from '../../util/colors';
+
+export const styles = StyleSheet.create({
+  textField: {
+    borderWidth: 1,
+    borderColor: colors.lightGray,
+    marginBottom: 15,
+    paddingLeft: 12,
+  },
+});
