+++ src/util/colors.js
@@ -1,0 +1,6 @@
+export const colors = {
+  blue: '#5c6ef8',
+  lightGray: '#4d4d4d',
+  white: '#fff',
+  black: '#000',
+};
